requirements.txt
jionlp
sentence-transformers
langchain
langchain-community
typing
presidio_analyzer
presidio_anonymizer